+++ src/Button.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { PropTypes } from 'react';
 import classNames from 'classnames';
 import LoadingSpinner from './LoadingSpinner';
 import omit from 'lodash/omit';
@@ -7,11 +7,11 @@ import omit from 'lodash/omit';
 var Button = React.createClass({
 
   propTypes: {
-    btnStyle: React.PropTypes.string,
-    block: React.PropTypes.bool,
-    processing: React.PropTypes.bool,
-    loadingSpinnerColor: React.PropTypes.string,
-    loadingSpinnerBackgroundColor: React.PropTypes.string,
+    btnStyle: PropTypes.string,
+    block: PropTypes.bool,
+    processing: PropTypes.bool,
+    loadingSpinnerColor: PropTypes.string,
+    loadingSpinnerBackgroundColor: PropTypes.string
   },
 
   render: function() {
@@ -24,7 +24,7 @@ var Button = React.createClass({
         're-btn--danger': btnStyle === 'danger',
         're-btn--link': btnStyle === 'link',
         're-btn--block': this.props.block,
-        're-btn--processing': this.props.processing,
+        're-btn--processing': this.props.processing
       },
       this.props.className
     );
@@ -53,5 +53,6 @@ var Button = React.createClass({
   }
 
 });
+
 
 module.exports = Button;